requirements.txt
flask
flask-restx
flask-sqlalchemy
werkzeug
python-dotenv
